+++ cpu-app/ClientApp/src/app/app-routing.module.ts
@@ -2,13 +2,13 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { HomeComponent } from './home/home.component';
 import { RenewApplicationComponent } from './renew-application/renew-application.component';
-import { NotFoundComponent } from './pages/not-found/not-found.component';
 import { OrganizationProfilePageComponent } from './pages/organization-profile-page/organization-profile-page.component';
 import { DashboardPageComponent } from './pages/dashboard-page/dashboard-page.component';
 import { ProgramPageComponent } from './pages/program-page/program-page.component';
 import { ExpensePageComponent } from './pages/expense-page/expense-page.component';
 import { BudgetPageComponent } from './pages/budget-page/budget-page.component';
 import { StatusReportPageComponent } from './pages/status-report-page/status-report-page.component';
+import { NotFoundPageComponent } from './pages/not-found-page/not-found-page.component';
 
 const routes: Routes = [
   {
@@ -45,7 +45,7 @@ const routes: Routes = [
     component: OrganizationProfilePageComponent,
     //canDeactivate: [CanDeactivateGuard]
   },
-  { path: '**', component: NotFoundComponent }
+  { path: '**', component: NotFoundPageComponent }
 ];
 
 @NgModule({
