+++ cpu-app/ClientApp/src/app/classes/readme.md
@@ -1,0 +1,1 @@
+These are actually models. However they don't parallel the back end models. They should be moved into the models folder when the Dynamics cruft is cleaned out of there.
