+++ cpu-app/ClientApp/src/app/pages/not-found-page/not-found-page.component.ts
@@ -2,15 +2,15 @@ import { Component } from '@angular/core';
 import { Router } from "@angular/router";
 
 @Component({
-    selector: 'app-not-found',
-    templateUrl: './not-found.component.html',
-    styleUrls: ['./not-found.component.scss']
+  selector: 'app-not-found-page',
+  templateUrl: './not-found-page.component.html',
+  styleUrls: ['./not-found-page.component.scss']
 })
 /** NotFound component*/
-export class NotFoundComponent {
-    /** NotFound ctor */
+export class NotFoundPageComponent {
+  /** NotFound ctor */
   constructor(private router: Router) {
     this.router.navigateByUrl("/404");
 
-    }
+  }
 }
