+++ cpu-app/ClientApp/src/app/shared/components/cancel-dialog/cancel-dialog.component.html
