+++ cpu-app/ClientApp/src/app/shared/components/field/field.component.ts
