+++ cpu-app/ClientApp/src/app/shared/components/file-uploader/file-uploader.component.scss
