+++ cpu-app/ClientApp/src/app/shared/components/guard-dialog/guard-dialog.component.ts
