+++ cpu-app/ClientApp/src/app/shared/components/quick-exit/quick-exit.component.html
