+++ cpu-app/ClientApp/src/app/shared/components/sign-dialog/sign-dialog.component.ts
