+++ cpu-app/ClientApp/src/app/shared/components/version-info/version-info.component.scss
